Reshuffle data structures in targeted files and runtime module
