+++ src/lib.rs
@@ -8,4 +8,5 @@ mod policy;
 pub mod policy_evaluator;
 pub mod policy_metadata;
 mod policy_tracing;
+pub mod runtimes;
 pub mod validation_response;
