+++ src/policy_evaluator.rs
@@ -2,18 +2,18 @@ use anyhow::{anyhow, Result};
 use lazy_static::lazy_static;
 use serde::Serialize;
 use serde_json::{json, value};
-use std::{collections::HashMap, convert::TryFrom, fmt, fs, path::Path, sync::RwLock};
+use std::{collections::HashMap, fmt, fs, path::Path, sync::RwLock};
 use tracing::error;
 
 use wapc::WapcHost;
 use wasmtime_provider::WasmtimeEngineProvider;
 
 use kubewarden_policy_sdk::metadata::ProtocolVersion;
-use kubewarden_policy_sdk::response::ValidationResponse as PolicyValidationResponse;
 use kubewarden_policy_sdk::settings::SettingsValidationResponse;
 
 use crate::cluster_context::ClusterContext;
 use crate::policy::Policy;
+use crate::runtimes::{burrego::Runtime as BurregoRuntime, wapc::Runtime as WapcRuntime};
 use crate::validation_response::ValidationResponse;
 
 pub enum PolicyExecutionMode {
@@ -28,14 +28,14 @@ lazy_static! {
 }
 
 #[derive(Serialize)]
-pub struct ValidateRequest(serde_json::Value);
+pub struct ValidateRequest(pub(crate) serde_json::Value);
 
 impl ValidateRequest {
     pub fn new(request: serde_json::Value) -> Self {
         ValidateRequest(request)
     }
 
-    fn uid(&self) -> &str {
+    pub(crate) fn uid(&self) -> &str {
         if let Some(uid) = self.0.get("uid").and_then(value::Value::as_str) {
             uid
         } else {
@@ -97,202 +97,20 @@ pub(crate) fn host_callback(
     }
 }
 
-struct BurregoEvaluator {
-    evaluator: burrego::opa::wasm::Evaluator,
-    entrypoint_id: i32,
-    input: serde_json::Value,
-    data: serde_json::Value,
+pub(crate) struct BurregoEvaluator {
+    pub(crate) evaluator: burrego::opa::wasm::Evaluator,
+    pub(crate) entrypoint_id: i32,
+    pub(crate) input: serde_json::Value,
+    pub(crate) data: serde_json::Value,
 }
+
+pub(crate) type PolicySettings = serde_json::Map<String, serde_json::Value>;
 
 enum Runtime {
     Wapc(wapc::WapcHost),
     // The `BurregoEvaluator` variant is boxed since it outsizes the
     // other variants of this enum.
     Burrego(Box<BurregoEvaluator>),
-}
-
-struct WapcRuntime<'a>(&'a mut wapc::WapcHost);
-struct BurregoRuntime<'a>(&'a mut BurregoEvaluator);
-
-type PolicySettings = serde_json::Map<String, serde_json::Value>;
-
-impl<'a> WapcRuntime<'a> {
-    fn validate(
-        &mut self,
-        settings: &PolicySettings,
-        request: &ValidateRequest,
-    ) -> ValidationResponse {
-        let uid = request.uid();
-
-        let req_obj = match request.0.get("object") {
-            Some(req_obj) => req_obj,
-            None => {
-                return ValidationResponse::reject(
-                    uid.to_string(),
-                    "request doesn't have an 'object' value".to_string(),
-                    hyper::StatusCode::BAD_REQUEST.as_u16(),
-                );
-            }
-        };
-
-        let validate_params = json!({
-            "request": request,
-            "settings": settings,
-        });
-
-        let validate_str = match serde_json::to_string(&validate_params) {
-            Ok(s) => s,
-            Err(e) => {
-                error!(
-                    error = e.to_string().as_str(),
-                    "cannot serialize validation params"
-                );
-                return ValidationResponse::reject_internal_server_error(
-                    uid.to_string(),
-                    e.to_string(),
-                );
-            }
-        };
-
-        match self.0.call("validate", validate_str.as_bytes()) {
-            Ok(res) => {
-                let pol_val_resp: Result<PolicyValidationResponse> = serde_json::from_slice(&res)
-                    .map_err(|e| anyhow!("cannot deserialize policy validation response: {:?}", e));
-                pol_val_resp
-                    .and_then(|pol_val_resp| {
-                        ValidationResponse::from_policy_validation_response(
-                            uid.to_string(),
-                            req_obj,
-                            &pol_val_resp,
-                        )
-                    })
-                    .unwrap_or_else(|e| {
-                        error!(
-                            error = e.to_string().as_str(),
-                            "cannot build validation response from policy result"
-                        );
-                        ValidationResponse::reject_internal_server_error(
-                            uid.to_string(),
-                            e.to_string(),
-                        )
-                    })
-            }
-            Err(e) => {
-                error!(error = e.to_string().as_str(), "waPC communication error");
-                ValidationResponse::reject_internal_server_error(uid.to_string(), e.to_string())
-            }
-        }
-    }
-
-    fn validate_settings(&mut self, settings: String) -> SettingsValidationResponse {
-        match self.0.call("validate_settings", settings.as_bytes()) {
-            Ok(res) => {
-                let vr: Result<SettingsValidationResponse> = serde_json::from_slice(&res)
-                    .map_err(|e| anyhow!("cannot convert response: {:?}", e));
-                vr.unwrap_or_else(|e| SettingsValidationResponse {
-                    valid: false,
-                    message: Some(format!("error: {:?}", e)),
-                })
-            }
-            Err(err) => SettingsValidationResponse {
-                valid: false,
-                message: Some(format!(
-                    "Error invoking settings validation callback: {:?}",
-                    err
-                )),
-            },
-        }
-    }
-
-    fn protocol_version(&self) -> Result<ProtocolVersion> {
-        match self.0.call("protocol_version", &[0; 0]) {
-            Ok(res) => ProtocolVersion::try_from(res.clone()).map_err(|e| {
-                anyhow!(
-                    "Cannot create ProtocolVersion object from '{:?}': {:?}",
-                    res,
-                    e
-                )
-            }),
-            Err(err) => Err(anyhow!(
-                "Cannot invoke 'protocol_version' waPC function: {:?}",
-                err
-            )),
-        }
-    }
-}
-
-impl<'a> BurregoRuntime<'a> {
-    fn validate(
-        &mut self,
-        _settings: &PolicySettings,
-        request: &ValidateRequest,
-    ) -> ValidationResponse {
-        let uid = request.uid();
-
-        let burrego_evaluation =
-            self.0
-                .evaluator
-                .evaluate(self.0.entrypoint_id, &self.0.input, &self.0.data);
-
-        match burrego_evaluation {
-            Ok(evaluation_result) => {
-                // According to the OPA WASM
-                // documentation/specification, the result of
-                // the evaluation will contain an object of
-                // the form:
-                //
-                // ```
-                // [
-                //   {
-                //     "result": "<the AdmissionReview object -- generated by the policy>"
-                //   }
-                // ]
-                // ```
-                //
-                // Reference: https://www.openpolicyagent.org/docs/v0.31.0/wasm/#compiling-policies
-                let evaluation_result = evaluation_result
-                    .get(0)
-                    .and_then(|r| r.get("result"))
-                    .and_then(|r| r.get("response"));
-
-                match evaluation_result {
-                    Some(evaluation_result) => {
-                        match serde_json::from_value(evaluation_result.clone()) {
-                            Ok(evaluation_result) => ValidationResponse {
-                                uid: uid.to_string(),
-                                ..evaluation_result
-                            },
-                            Err(err) => ValidationResponse::reject_internal_server_error(
-                                uid.to_string(),
-                                err.to_string(),
-                            ),
-                        }
-                    }
-                    None => ValidationResponse::reject_internal_server_error(
-                        uid.to_string(),
-                        "cannot interpret OPA policy result".to_string(),
-                    ),
-                }
-            }
-            Err(err) => {
-                error!(
-                    error = err.to_string().as_str(),
-                    "error evaluating policy with burrego"
-                );
-                ValidationResponse::reject_internal_server_error(uid.to_string(), err.to_string())
-            }
-        }
-    }
-
-    fn validate_settings(&mut self, _settings: String) -> SettingsValidationResponse {
-        // The burrego backend is mainly for compatibility with
-        // existing OPA policies. Those policies don't have a generic
-        // way of validating settings. Return true
-        SettingsValidationResponse {
-            valid: true,
-            message: None,
-        }
-    }
 }
 
 pub struct PolicyEvaluator {
@@ -404,7 +222,7 @@ impl PolicyEvaluator {
             Err(err) => {
                 return SettingsValidationResponse {
                     valid: false,
-                    message: Some(format!("could not marshal setings: {}", err)),
+                    message: Some(format!("could not marshal settings: {}", err)),
                 }
             }
         };
