+++ src/runtime.rs
@@ -1,0 +1,6 @@
+enum Runtime {
+    Wapc(wapc::WapcHost),
+    // The `BurregoEvaluator` variant is boxed since it outsizes the
+    // other variants of this enum.
+    Burrego(Box<BurregoEvaluator>),
+}
